Move WaterControl to dedicated irrigation.py module
